Task_4: created dropdown component
