+++ src/components/Post/Post.scss
@@ -22,68 +22,6 @@
   &:hover > .hov {
     display: inline-block;
     text-align: right;
-    &:focus ~ .dropdown {
-      opacity: 1;
-      background: $darkGray;
-      width: 50%;
-      position: absolute;
-      right: 15px;
-    
-      ul {
-        padding: 0;
-      }
-    }
-  }
-
-  .btn-round {
-    position: relative;
-    border-radius: 50%;
-    border: none;
-    width: 36px;
-    height: 36px;
-    color: #2A202D;
-    background: #2A202D;
-    box-shadow: 0px 0px 4px rgba(0, 0, 0, 0.196596);
-  
-    &::after {
-      color: $white;
-      position: absolute;
-      right: 8px;
-      top: 4px;
-      text-align: center;
-      content: '\2807';
-      font-size: 20px;
-    }
-  }
-
-  .dropdown-btn {
-    @include text(16px, 500, 20px);
-    color: $white;
-    background: $darkGray;
-    border: none;
-    width: 100%;
-    text-align: left;
-    padding: 16px 23px;
-  
-    &:hover {
-      background:  $red;
-    }
-  }
-
-  .dropdown-btn-x {
-    @include text(16px, 500, 20px);
-    color: $white;
-    background: $darkGray;
-    border: none;
-    padding: 1px 2px;
-  
-    &:hover {
-      background:  $red;
-    }
-  }
-
-  .dropdown {
-    opacity: 0;
   }
 }
