+++ src/components/Post/Post.tsx
@@ -1,6 +1,6 @@
-import React, { useContext } from 'react';
-import { AppContext } from '../../Context';
+import React from 'react';
 import { Post as IPost } from '../../types/interfaces';
+import Dropdown from '../Dropdown/Dropdown';
 
 import './Post.scss';
 
@@ -9,47 +9,16 @@ interface PostItemProp {
 }
 
 const Post = ({ post }: PostItemProp) => {
-  const { toggleOpen, toggleModal, setEditedPostVal } = useContext(AppContext);
   const {
-    id, name, year, genre, imgName, rating, runtime,
+    name, year, genre,
   } = post;
-
-  const handleDeleteClick = () => {
-    toggleModal('delete');
-    toggleOpen(true);
-  };
-
-  const handleEditClick = () => {
-    toggleModal('edit');
-    toggleOpen(true);
-    setEditedPostVal({
-      id,
-      name,
-      year,
-      genre,
-      imgName,
-      rating,
-      runtime,
-    });
-  };
 
   return (
     <article className="movie">
       <div
         className="img"
       >
-        <button type="button" className="btn-round hov">.</button>
-        <div className="dropdown">
-          <button type="button" className="dropdown-btn-x">X</button>
-          <ul>
-            <li>
-              <button type="button" className="dropdown-btn" onClick={handleEditClick}>Edit</button>
-            </li>
-            <li>
-              <button type="button" className="dropdown-btn" onClick={handleDeleteClick}>Delete</button>
-            </li>
-          </ul>
-        </div>
+        <Dropdown post={post} />
       </div>
       <div className="description">
         <div className="name">
